fix(api): resolve critical crashes and security issues
- patterns.py: replace sync .query() with async select/execute; add
  missing await on db.commit() and db.delete()
- celery_tasks.py: fix sync session from async connection; re-raise
  exceptions so Celery retry engages
- auth.py: replace blocking requests.post with httpx.AsyncClient
- config.py: remove hardcoded JWT secret fallback; remove CORS wildcard
- schemas/user.py: remove tokens from UserResponse to prevent leakage
- transaction_service.py: fix O(n) count with func.count(); fix bulk
  update count for transactor_label-only updates
- transaction_schemas.py: amount int→float; add max_length validation
- accounts.py: cap pagination offset to prevent full-table scan
- dependencies.py: re-raise HTTPException before catch-all

+++ app/utils/transaction_serializer.py
@@ -18,7 +18,7 @@ def serialize_transaction(transaction: Transaction) -> Dict[str, Any]:
     """
     return {
         "id": transaction.id,
-        "amount": int(transaction.amount) if transaction.amount is not None else None,
+        "amount": float(transaction.amount) if transaction.amount is not None else None,
         "transaction_id": transaction.transaction_id,
         "type": transaction.type,
         "date": transaction.date.isoformat() if transaction.date else None,
